# openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/sign_up/":{"post":{"summary":"Create User","description":"User registration","operationId":"create_user_sign_up__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/sign_in/":{"post":{"summary":"Sign In User","description":"User sign in","operationId":"sign_in_user_sign_in__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignIn"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/my_user_info/":{"get":{"summary":"Get User Info","description":"Get my user information","operationId":"get_user_info_my_user_info__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"security":[{"JWTBearer":[]}]}},"/available_models/":{"get":{"summary":"Get Available Models","description":"Get my user information","operationId":"get_available_models_available_models__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/MlModel"},"type":"array","title":"Response Get Available Models Available Models  Get"}}}}},"security":[{"JWTBearer":[]}]}},"/infer_model/":{"post":{"summary":"Infer Model","description":"Create ML model inference request","operationId":"infer_model_infer_model__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InferenceRequestCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InferenceRequest"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"JWTBearer":[]}]}},"/inf_requests/":{"get":{"summary":"Get Inf Requests","description":"Get inference requests information","operationId":"get_inf_requests_inf_requests__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/InferenceRequest"},"type":"array","title":"Response Get Inf Requests Inf Requests  Get"}}}}},"security":[{"JWTBearer":[]}]}},"/inf_result/{request_id}":{"get":{"summary":"Get Inf Result","description":"Get inference result by inference request id","operationId":"get_inf_result_inf_result__request_id__get","security":[{"JWTBearer":[]}],"parameters":[{"name":"request_id","in":"path","required":true,"schema":{"type":"integer","title":"Request Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InferenceResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"InferenceRequest":{"properties":{"SEQN":{"type":"number","title":"Seqn"},"RIAGENDR":{"type":"number","title":"Riagendr"},"PAQ605":{"type":"number","title":"Paq605"},"BMXBMI":{"type":"number","title":"Bmxbmi"},"LBXGLU":{"type":"number","title":"Lbxglu"},"DIQ010":{"type":"number","title":"Diq010"},"LBXGLT":{"type":"number","title":"Lbxglt"},"LBXIN":{"type":"number","title":"Lbxin"},"id":{"type":"integer","title":"Id"},"status":{"type":"integer","title":"Status"},"cost":{"type":"integer","title":"Cost"},"inference_result":{"anyOf":[{"$ref":"#/components/schemas/InferenceResult"},{"type":"null"}]},"model":{"$ref":"#/components/schemas/MlModel"}},"type":"object","required":["SEQN","RIAGENDR","PAQ605","BMXBMI","LBXGLU","DIQ010","LBXGLT","LBXIN","id","status","cost","inference_result","model"],"title":"InferenceRequest"},"InferenceRequestCreate":{"properties":{"SEQN":{"type":"number","title":"Seqn"},"RIAGENDR":{"type":"number","title":"Riagendr"},"PAQ605":{"type":"number","title":"Paq605"},"BMXBMI":{"type":"number","title":"Bmxbmi"},"LBXGLU":{"type":"number","title":"Lbxglu"},"DIQ010":{"type":"number","title":"Diq010"},"LBXGLT":{"type":"number","title":"Lbxglt"},"LBXIN":{"type":"number","title":"Lbxin"},"ml_model_id":{"type":"integer","title":"Ml Model Id"}},"type":"object","required":["SEQN","RIAGENDR","PAQ605","BMXBMI","LBXGLU","DIQ010","LBXGLT","LBXIN","ml_model_id"],"title":"InferenceRequestCreate"},"InferenceResult":{"properties":{"id":{"type":"integer","title":"Id"},"value":{"type":"string","title":"Value"}},"type":"object","required":["id","value"],"title":"InferenceResult"},"MlModel":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"price":{"type":"integer","title":"Price"},"description":{"type":"string","title":"Description"}},"type":"object","required":["id","name","price","description"],"title":"MlModel"},"SignIn":{"properties":{"email":{"type":"string","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"SignIn"},"SignInResult":{"properties":{"access_token":{"type":"string","title":"Access Token"},"expiration":{"type":"string","format":"date-time","title":"Expiration"},"user_info":{"$ref":"#/components/schemas/User"}},"type":"object","required":["access_token","expiration","user_info"],"title":"SignInResult"},"User":{"properties":{"email":{"type":"string","title":"Email"},"id":{"type":"integer","title":"Id"},"is_active":{"type":"boolean","title":"Is Active"},"is_superuser":{"type":"boolean","title":"Is Superuser"},"balance":{"type":"integer","title":"Balance"}},"type":"object","required":["email","id","is_active","is_superuser","balance"],"title":"User"},"UserCreate":{"properties":{"email":{"type":"string","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"UserCreate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"JWTBearer":{"type":"http","scheme":"bearer"}}}}